fix: move services and repositories to web package
