+++ src/main/java/org/storkforge/barkr/web/controller/WebController.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import org.storkforge.barkr.domain.AccountService;
-import org.storkforge.barkr.domain.PostService;
+import org.storkforge.barkr.web.domain.AccountService;
+import org.storkforge.barkr.web.domain.PostService;
 import org.storkforge.barkr.dto.accountDto.ResponseAccount;
 import org.storkforge.barkr.dto.postDto.CreatePost;
