+++ src/main/java/org/storkforge/barkr/web/domain/PostService.java
@@ -1,4 +1,4 @@
-package org.storkforge.barkr.domain;
+package org.storkforge.barkr.web.domain;
 
 import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
@@ -8,8 +8,8 @@ import org.storkforge.barkr.domain.entity.Account;
 import org.storkforge.barkr.domain.entity.Post;
 import org.storkforge.barkr.dto.postDto.CreatePost;
 import org.storkforge.barkr.dto.postDto.ResponsePost;
-import org.storkforge.barkr.infrastructure.persistence.AccountRepository;
-import org.storkforge.barkr.infrastructure.persistence.PostRepository;
+import org.storkforge.barkr.web.infrastructure.persistence.AccountRepository;
+import org.storkforge.barkr.web.infrastructure.persistence.PostRepository;
 import org.storkforge.barkr.mapper.PostMapper;
 
 import java.util.List;
